+++ demo/symfony8/config/bundles.php
@@ -4,12 +4,14 @@ declare(strict_types=1);
 
 use Doctrine\Bundle\DoctrineBundle\DoctrineBundle;
 use Nowo\ApiStudioBundle\ApiStudioBundle;
+use Nowo\FormKitBundle\NowoFormKitBundle;
 use Nowo\TwigInspectorBundle\NowoTwigInspectorBundle;
 use Nowo\UiKitBundle\NowoUiKitBundle;
 use Symfony\Bundle\FrameworkBundle\FrameworkBundle;
 use Symfony\Bundle\SecurityBundle\SecurityBundle;
 use Symfony\Bundle\TwigBundle\TwigBundle;
 use Symfony\Bundle\WebProfilerBundle\WebProfilerBundle;
+use Twig\Extra\TwigExtraBundle\TwigExtraBundle;
 
 return [
     FrameworkBundle::class         => ['all' => true],
@@ -20,4 +22,6 @@ return [
     ApiStudioBundle::class         => ['all' => true],
     NowoUiKitBundle::class         => ['all' => true],
     NowoTwigInspectorBundle::class => ['dev' => true, 'test' => true],
+    TwigExtraBundle::class         => ['all' => true],
+    NowoFormKitBundle::class       => ['all' => true],
 ];